Blues: Use PNG instead of css filter for gray logo
Using the CSS filter is unreliable. It is ugly in IE
and doesn't work at all in older versions of Safari.

So, use a PNG instead, and set it up using jQuery.

Also, because the logo is inside the RSVP form, which
is inside the Angular view, we have to do a little
finessing to make sure the jQuery is executed at
the proper time.