Support dev on Mac OS X
Support development using Apache and Mono,
while also working on IIS.

There are a few things that need to be changed
in order to work with Mono -- mainly setting
VB as the default language.

With Apache, I couldn't figure out how to get
ASP.NET code working in the .html files (used
in SSI), so I renamed the .html include files
to end with .aspx.